hirtectl: moved list-units to dedicated src file
- moved list-unit method incl. structs to dedicated src file
- replaced printer struct by simple function pointer
