+++ src/client/types.h
@@ -3,4 +3,3 @@
 
 typedef struct Client Client;
 typedef struct UnitList UnitList;
-typedef struct Printer Printer;
